# version/server/Server/api/openapi.yaml
openapi: 3.0.1
info:
  title: Swagger ToDo Manager
  description: REST APIs for the interaction with the ToDo Manager service.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: https://todo.swagger.io/v2
- url: http://todo.swagger.io/v2
...
paths:
...
  /api/tasks/{taskId}/assignees:
    get:
      tags:
      - "tasks/{taskId}/assignees"
      summary: Retreve the users a task has been assigned to
      description: The users to whom the task with ID taskId is assigned are retrieved.
        This operation can be performed only by the owner.
      operationId: getUsersAssigned
      parameters:
      - name: taskId
        in: path
        description: ID of the task to retrieve
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
                x-content-type: application/json
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskIdassignees
    post:
      tags:
      - "tasks/{taskId}/assignees"
      summary: Assign a task to a user
      description: The task with ID taskId is assigned to the user specified in the
        request body. This operation can only be performed by the owner.
      operationId: assignTaskToUser
      parameters:
      - name: taskId
        in: path
        description: ID of the task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        description: the user to whom the task is assigned
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "400":
          description: Bad Request
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskIdassignees
  /api/tasks/{taskId}/assignees/{userId}:
    delete:
      tags:
      - "tasks/{taskId}/assignees/{userId}"
      summary: Remove a user from the assigned task
      description: "The user that is removed is the user, identified by userId, that was assigned to the task identified by taskId. This operation can be performed only by the owner."
      operationId: removeUser
      parameters:
      - name: taskId
        in: path
        description: ID of the assigned task
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      - name: userId
        in: path
        description: ID of the user to remove
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Successful operation - No Content
          content: {}
        "401":
          description: Unauthorized
          content: {}
        "403":
          description: Forbidden
          content: {}
        "404":
          description: Not Found
          content: {}
        "500":
          description: Internal Server Error
          content: {}
      security:
      - cookieAuth: []
      x-swagger-router-controller: TaskstaskIdassigneesuserId
...
  '/api/tasks/{taskId}/images':
      post:
        tags:
          - tasks/{taskId}/images
        security:
          - cookieAuth: []
        summary: Associate an image to the task
        description: 
          The image sent in request body is associated to the task characterized by the ID specified in the path. Only the task owner can associate an image to the task.
        operationId: addImage
        parameters:
          - name: taskId
            in: path
            description: ID of the task
            required: true
            schema:
              type: integer
              format: int64
        requestBody:
          description: image to assign to the task
          content:
            multipart/form-data:
              schema:
                type: object
                properties: # Request parts
                  image:
                      type: string
                      format: binary
              encoding: 
                  image: 
                    contentType: image/png, image/jpg, image/gif
          required: true
        responses:
          '201':
            description: Created
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Image'
          '403':
            description: Forbidden
            content: {}
          '404':
            description: Not Found
            content: {}
          '415':
            description: Unsupported Media Type
            content: {}
          '500':
            description: Internal Server Error
            content: {}
  '/api/tasks/{taskId}/img/{imageId}':
      get:
        tags:
        - "tasks/{taskId}/image/{imageId}"
        summary: Retrieve the image associated to the task with ID imgId
        description: This operation return the image associated to the task with a specified img id."
        operationId: getImage
        parameters:
        - name: taskId
          in: path
          description: ID of the user
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: "The id of the requested image"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        responses:
          "200":
            description: Successful operation
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Image'
          "404":
            description: Not Found
            content: {}
          "500":
            description: Internal Server Error
            content: {}
        security:
          - cookieAuth: []
      delete:
        tags:
        - "tasks/{taskId}/image/{imageId}"
        summary: Delete the image associated to the task with ID imgId
        description: This operation return the image associated to the task with a specified img id."
        operationId: deleteImage
        parameters:
        - name: taskId
          in: path
          description: ID of the user
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: "The id of the requested image"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        responses:
          '204':
            description: No Content
            content: {}
          '403':
            description: Forbidden
            content: {}
          '404':
            description: Not Found
            content: {}
          '500':
            description: Internal Server Error
            content: {}
        security:
          - cookieAuth: []
  '/api/tasks/{taskId}/images/{imageId}/imageFile':
    get:
      tags:
        - tasks/{taskId}/images/{imageId}/imageFile
      security:
        - cookieAuth: []
      summary: Retrieve the file of an image associated to the task
      description: 
        The image file to be retrieved is linked to the image having the ID specified in the path and it is associated to the task characterized by the ID specified in the path. Only the task owner or a task assignee can retrieve the image file.
      operationId: getFileImage
      parameters:
        - name: taskId
          in: path
          description: ID of the task
          required: true
          schema:
            type: integer
            format: int64
        - name: imageId
          in: path
          description: ID of the image
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpg:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '403':
          description: Forbidden
          content: {}
        '404':
          description: Not Found
          content: {}
        '414':
          description: Unsupported Media Type
          content: {}
        '500':
          description: Internal Server Error
          content: {}       
components:
  schemas:
...
    User:
      type: object
      properties:
        $schema:
          type: string
          description: path to the JSON Schema to use for the validation of the JSON
            file describing the task
        id:
          type: integer
          description: unique identifier for the user
        name:
          type: string
          description: username of the user
        email:
          type: string
          description: email address of the user
          format: email
        password:
          maxLength: 20
          minLength: 6
          type: string
          description: password of the user
        self:
          type: string
          description: self link
          format: uri-reference
      example:
        password: password
        $schema: $schema
        name: name
        self: self
        id: 5
        email: ""
...
    Image:
      type: object
      properties:
        $schema:
          type: string
          description: 
            path to the JSON Schema to use for the validation of the JSON file
            describing the image
        id:
          type: integer
          description: unique identifier for the image
        name:
          type: string
          description: name without extension of the image
        fileURI:
          type: string
          format: uri-reference
          description: link to the content of the image
        self:
          type: string
          format: uri-reference
          description: self link
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT to identify the authenticated user.
      name: connect.sid
      in: cookie